Add pharmacy costing service and refactor controllers
